requirements.txt
fastapi
uvicorn
paho-mqtt
pandas
requests
python-dotenv
streamlit
